refactor: use own logging interface
