+++ src/middleware/logger.middleware.ts
@@ -1,16 +1,14 @@
-import {
-  Inject,
-  Injectable,
-  LoggerService,
-  NestMiddleware,
-} from '@nestjs/common';
+import { Inject, Injectable, NestMiddleware } from '@nestjs/common';
 import { NextFunction, Request, Response } from 'express';
-import { LoggingService } from '../routes/common/logging/logging.service';
+import {
+  ILoggingService,
+  LoggingService,
+} from '../routes/common/logging/logging.interface';
 
 @Injectable()
 export class LoggerMiddleware implements NestMiddleware {
   constructor(
-    @Inject(LoggingService) private readonly loggingService: LoggerService,
+    @Inject(LoggingService) private readonly loggingService: ILoggingService,
   ) {}
 
   use(req: Request, res: Response, next: NextFunction) {
