+++ src/routes/common/logging/logging.service.ts
@@ -1,11 +1,15 @@
-import { Injectable, LoggerService } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { ClsService } from 'nestjs-cls';
 import * as winston from 'winston';
+import { ILoggingService } from './logging.interface';
 
-export const LoggingService = Symbol('LoggerService');
-
+/**
+ * Implementation of LoggerService which prepends the current time and a unique request ID to every logged message.
+ *
+ * The requestID is generated and provided using the `nestjs-cls` `ClsService` which uses the async local storage to store a uuid for each processed request through a middleware.
+ */
 @Injectable()
-export class RequestScopedLoggingService implements LoggerService {
+export class RequestScopedLoggingService implements ILoggingService {
   constructor(private readonly cls: ClsService) {}
   log(message: string, ...optionalParams: any[]) {
     winston.info(this.transformMessage(message), ...optionalParams);
@@ -21,9 +25,6 @@ export class RequestScopedLoggingService implements LoggerService {
   }
   verbose(message: string, ...optionalParams: any[]) {
     winston.verbose(this.transformMessage(message), ...optionalParams);
-  }
-  setLogLevels() {
-    // We do not use this
   }
 
   private transformMessage(message: string): string {
